do_unpack: utilize the python compression modules for the unpack process
Sadly, this slows down a task-boot do_unpack_all from 2m30s to 3m, massive
percentage drop in performance.  A comparison of the unpack of
gcc-4.4.2.tar.bz2 by the tar command vs python shows one that the python
implementation is half the speed, so this is likely to remain just a proof of
concept.

Signed-off-by: Chris Larson <chris_larson@mentor.com>