base, rm_work: make workdir cleanup opt-out
Rather than requiring the user to INHERIT += "rm_work" to get their workdirs
cleaned up during a build, that is now default behavior, and you may set
RM_WORK = "" to disable this behavior.

This also switches the actual workdir cleanup to occur inside the do_build
task rather than via 2 new tasks.  As a consequence of changing the method in
this way, certain previous behaviors will behave a bit differently.

As one example, adding a new INHERIT of a package_* class after a recipe has
already been fully built will result in having to rebuild the recipe, as its
workdir will have been cleaned.

As another example, changing RM_WORK will only affect the next recipes to be
built, not the previous ones, so changing it mid-build will not behave the
same as adding INHERIT of rm_work mid-build behaves today.

I personally feel that the simplicity gained and the reduction in number of
tasks are worthwhile improvements and outweigh the loss of what I believe are
relatively rarely used features.

Signed-off-by: Chris Larson <chris_larson@mentor.com>